pyproject.toml
[project]
name = "sass-billing"
version = "0.1.0"
description = ""
authors = [
    {name = "hafis",email = "muhammed.hafis@simelabs.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-guardian (>=3.0.0,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "xhtml2pdf (>=0.2.17,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
